requirements.txt
pytest==8.3.2
selenium==4.24.0
allure-pytest==2.13.5
requests==2.32.3
allure-python-commons==2.13.5